% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtestbayesianWrapper.R
\name{ABTestBayesian}
\alias{ABTestBayesian}
\title{Bayesian A/B Test}
\usage{
ABTestBayesian(
  data = NULL,
  version = "1",
  bayesFactorOrder = "bestModelTop",
  bayesFactorType = "BF10",
  bfRobustnessPlot = FALSE,
  bfRobustnessPlotLowerPriorMean = -0.5,
  bfRobustnessPlotLowerPriorSd = 0.1,
  bfRobustnessPlotStepsPriorMean = 5,
  bfRobustnessPlotStepsPriorSd = 5,
  bfRobustnessPlotType = "BF10",
  bfRobustnessPlotUpperPriorMean = 0.5,
  bfRobustnessPlotUpperPriorSd = 1,
  bfSequentialPlot = FALSE,
  descriptivesTable = FALSE,
  n1 = list(types = list(), value = ""),
  n2 = list(types = list(), value = ""),
  normalPriorMean = 0,
  normalPriorSd = 1,
  plotHeight = 320,
  plotWidth = 480,
  priorModelProbabilityEqual = 0.5,
  priorModelProbabilityGreater = 0.25,
  priorModelProbabilityLess = 0.25,
  priorModelProbabilityTwoSided = 0,
  priorPlot = FALSE,
  priorPlotType = "logOddsRatio",
  priorPosteriorPlot = FALSE,
  priorPosteriorPlotType = "logOddsRatio",
  samples = 10000,
  seed = 1,
  setSeed = FALSE,
  y1 = list(types = list(), value = ""),
  y2 = list(types = list(), value = "")
)
}
\arguments{
\item{bayesFactorOrder, }{Compares each model against the model selected
\itemize{
  \item \code{"bestModelTop"}
  \item \code{"nullModelTop"}
}}

\item{bfRobustnessPlot, }{Displays the prior sensitivity analysis.
Defaults to \code{FALSE}.}

\item{bfRobustnessPlotStepsPriorMean, }{Specifies in how many discrete steps the μ step range is partitioned.}

\item{bfRobustnessPlotStepsPriorSd, }{Specifies in how many discrete steps the σ step range is partitioned.}

\item{bfSequentialPlot, }{Displays the development of posterior probabilities as the data come in. The probability wheels visualize prior and posterior probabilities of the hypotheses.
Defaults to \code{FALSE}.}

\item{descriptivesTable, }{Displays the counts and proportion for each group.
Defaults to \code{FALSE}.}

\item{n1, }{Number of trials in group 1 (control condition).}

\item{n2, }{Number of trials in group 2 (experimental condition).}

\item{normalPriorMean, }{Specifies the mean for the normal prior on the test-relevant log odds ratio.}

\item{normalPriorSd, }{Specifies the standard deviation for the normal prior on the test-relevant log odds ratio.}

\item{priorModelProbabilityEqual, }{Specifies that the 'success' probability is identical (there is no effect).}

\item{priorModelProbabilityGreater, }{Specifies that the 'success' probability in the experimental condition is higher than in the control condition.}

\item{priorModelProbabilityLess, }{Specifies that the 'success' probability in the experimental condition is lower than in the control condition.}

\item{priorModelProbabilityTwoSided, }{Specifies that the 'success' probability differs between the control and experimental condition, but does not specify which one is higher.}

\item{priorPlot, }{Plots parameter prior distributions.
Defaults to \code{FALSE}.}

\item{priorPosteriorPlot, }{Displays the prior and posterior density for the quantity of interest.
Defaults to \code{FALSE}.}

\item{samples, }{Specifies the number of importance samples for obtaining log marginal likelihood for (H+) and (H-) and the number of posterior samples.}

\item{y1, }{Number of successes in group 1 (control condition).}

\item{y2, }{Number of successes in group 2 (experimental condition).}
}
\description{
The Bayesian A/B test allows one to monitor the evidence for the hypotheses that an intervention or treatment has either a positive effect, a negative effect or no effect.
}
