% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialtestWrapper.R
\name{BinomialTest}
\alias{BinomialTest}
\title{Binomial Test}
\usage{
BinomialTest(
  data = NULL,
  version = "1",
  formula = NULL,
  alternative = "twoSided",
  ci = FALSE,
  ciLevel = 0.95,
  counts = list(types = list(), value = ""),
  descriptivesPlot = FALSE,
  descriptivesPlotCiLevel = 0.95,
  plotHeight = 320,
  plotWidth = 480,
  testValue = 0.5,
  variables = list(types = list(), value = list()),
  vovkSellke = FALSE
)
}
\arguments{
\item{ci, }{Coverage of the confidence intervals in percentages. The default value is 95.
Defaults to \code{FALSE}.}

\item{descriptivesPlot, }{Visualises the proportion and confidence interval of the two different values of your dichotomous variable.
Defaults to \code{FALSE}.}

\item{descriptivesPlotCiLevel, }{Coverage of the confidence intervals in percentages. The default value is 95.}

\item{testValue, }{The proportion of the variable under the null hypothesis - the baseline for comparison.}

\item{vovkSellke, }{An upper bound on how much more likely a p-value is under the alternative hypothesis than under the null.
Defaults to \code{FALSE}.}
}
\description{
The binomial test allows the user to test whether a proportion of a dichotomous variable is equal to a test value (=presumed population value). The analysis returns a binomial test for each level of the dependent variable against all other levels, so it will also work for variables with more than two levels.
## Assumptions
- The variable should be a dichotomous scale.
- Observations should be independent.
}
