% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialtestbayesianWrapper.R
\name{BinomialTestBayesian}
\alias{BinomialTestBayesian}
\title{Bayesian Binomial Test}
\usage{
BinomialTestBayesian(
  data = NULL,
  version = "1",
  formula = NULL,
  alternative = "twoSided",
  bayesFactorType = "BF10",
  bfSequentialPlot = FALSE,
  counts = list(types = list(), value = ""),
  descriptivesPlot = FALSE,
  descriptivesPlotCiLevel = 0.95,
  plotHeight = 320,
  plotWidth = 480,
  priorA = 1,
  priorB = 1,
  priorPosteriorPlot = FALSE,
  priorPosteriorPlotAdditionalInfo = TRUE,
  testValue = 0.5,
  variables = list(types = list(), value = list())
)
}
\arguments{
\item{bfSequentialPlot, }{Displays the development of the Bayes factor as the data come in using the user-defined prior.
Defaults to \code{FALSE}.}

\item{descriptivesPlot, }{Display descriptives plots.
Defaults to \code{FALSE}.}

\item{descriptivesPlotCiLevel, }{Display central credible intervals. A credible interval shows the probability that the true effect size lies within certain values. The default credible interval is set at 95%.}

\item{priorA, }{Sets how much prior belief you have in success. When a = b = 1, this corresponds to a uniform prior distribution.}

\item{priorB, }{Sets how much prior belief you have in failure. When a = b = 1, this corresponds to a uniform prior distribution.}

\item{priorPosteriorPlot, }{Displays the prior and posterior density of the population proportion under the alternative hypothesis.
Defaults to \code{FALSE}.}

\item{priorPosteriorPlotAdditionalInfo, }{Shows the Bayes factor using the chosen prior, a probability wheel showing evidence for each hypothesis, and the median with 95% credible interval of the effect size.
Defaults to \code{TRUE}.}

\item{testValue, }{The proportion of the variable under the null hypothesis - the baseline for comparison.}
}
\description{
The Bayesian binomial test allows you to test whether a proportion of a dichotomous variable is equal to a test value (presumed population value). The analysis returns a binomial test for each level of the dependent variable against all other levels, so it will also work for variables with more than two levels.
## Assumptions
- The variable should be a dichotomous scale.
- Observations should be independent.
}
