% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencytablesWrapper.R
\name{ContingencyTables}
\alias{ContingencyTables}
\title{Contingency Tables}
\usage{
ContingencyTables(
  data = NULL,
  version = "1",
  formula = NULL,
  byIntervalEta = FALSE,
  chiSquared = TRUE,
  chiSquaredContinuityCorrection = FALSE,
  cochranAndMantel = FALSE,
  columnOrder = "ascending",
  columns = list(types = list(), value = list()),
  contingencyCoefficient = FALSE,
  counts = list(types = list(), value = ""),
  countsExpected = FALSE,
  countsHiddenSmallCounts = FALSE,
  countsHiddenSmallCountsThreshold = 5,
  countsObserved = TRUE,
  gamma = FALSE,
  kendallsTauB = FALSE,
  kendallsTauC = FALSE,
  lambda = FALSE,
  layers = list(),
  likelihoodRatio = FALSE,
  marginShowTotals = TRUE,
  mcNemarChiSquared = FALSE,
  mcNemarChiSquaredContinuityCorrection = FALSE,
  oddsRatio = FALSE,
  oddsRatioAlternative = "twoSided",
  oddsRatioAsLogOdds = TRUE,
  oddsRatioCiLevel = 0.95,
  percentagesColumn = FALSE,
  percentagesRow = FALSE,
  percentagesTotal = FALSE,
  phiAndCramersV = FALSE,
  plotHeight = 320,
  plotWidth = 480,
  residualsPearson = FALSE,
  residualsStandardized = FALSE,
  residualsUnstandardized = FALSE,
  rowOrder = "ascending",
  rows = list(types = list(), value = list()),
  somersD = FALSE,
  testOddsRatioEquals = 1,
  uncertaintyCoefficient = FALSE,
  vovkSellke = FALSE,
  zTestAdjustedPValues = FALSE,
  zTestColumnComparison = FALSE
)
}
\arguments{
\item{chiSquared, }{Pearson’s chi-squared test for independence.
Defaults to \code{TRUE}.}

\item{chiSquaredContinuityCorrection, }{Applies Yates’ correction for continuity (for 2×2 tables).
Defaults to \code{FALSE}.}

\item{contingencyCoefficient, }{Measure of association for nominal variables (based on χ²).
Defaults to \code{FALSE}.}

\item{countsExpected, }{Show counts expected under the null hypothesis.
Defaults to \code{FALSE}.}

\item{countsObserved, }{Show actual counts from the data.
Defaults to \code{TRUE}.}

\item{gamma, }{Measure of ordinal association based on concordant/discordant pairs.
Defaults to \code{FALSE}.}

\item{kendallsTauB, }{Ordinal correlation adjusting for ties.
Defaults to \code{FALSE}.}

\item{lambda, }{Proportion reduction in error measure for nominal data.
Defaults to \code{FALSE}.}

\item{likelihoodRatio, }{Calculates the likelihood of the data under the alternative hypothesis divided by the likelihood of the data under the null hypothesis.
Defaults to \code{FALSE}.}

\item{marginShowTotals, }{Shows row and column totals.
Defaults to \code{TRUE}.}

\item{mcNemarChiSquared, }{Tests equality of two marginal proportions in paired nominal data.
Defaults to \code{FALSE}.}

\item{mcNemarChiSquaredContinuityCorrection, }{Corrects for error introduced when approximating a discrete distribution with a continuous one in McNemar's test.
Defaults to \code{FALSE}.}

\item{oddsRatio, }{Displays the odds ratio for 2×2 tables (ad/bc).
Defaults to \code{FALSE}.}

\item{oddsRatioAlternative, }{Specify the direction of the alternative hypothesis for Fisher's exact test.
\itemize{
  \item \code{"twoSided"}: Two-sided alternative hypothesis that the proportion of group 1 is not equal to the proportion of group 2.
  \item \code{"greater"}: One-sided alternative hypothesis that the proportion of group 1 is greater than the proportion of group 2.
  \item \code{"less"}: One-sided alternative hypothesis that the proportion of group 1 is less than the proportion of group 2.
}}

\item{oddsRatioAsLogOdds, }{Shows the log-transformed odds ratio.
Defaults to \code{TRUE}.}

\item{oddsRatioCiLevel, }{Coverage of the confidence intervals in percentages. The default value is 95.}

\item{percentagesColumn, }{Shows column-wise percentages.
Defaults to \code{FALSE}.}

\item{percentagesRow, }{Shows row-wise percentages.
Defaults to \code{FALSE}.}

\item{percentagesTotal, }{Shows percentages of total.
Defaults to \code{FALSE}.}

\item{phiAndCramersV, }{Effect size for nominal association; Phi (2×2), Cramer’s V (larger tables).
Defaults to \code{FALSE}.}

\item{residualsPearson, }{Standardized residuals; computed by (observed - expected) / √(expected).
Defaults to \code{FALSE}.}

\item{residualsStandardized, }{Accounts for row/column totals; computed by (observed - expected) / √(expected × (1 - row marginal proportion) × (1 - column marginal proportion)).
Defaults to \code{FALSE}.}

\item{residualsUnstandardized, }{Computed by (observed - expected).
Defaults to \code{FALSE}.}

\item{vovkSellke, }{An upper bound on how much more likely a p-value is under the alternative hypothesis than under the null.
Defaults to \code{FALSE}.}
}
\description{
Contingency tables allow the user to identify how the frequencies of one categorical variable relate to another, helping to determine associations between variables.
}
