% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencytablesbayesianWrapper.R
\name{ContingencyTablesBayesian}
\alias{ContingencyTablesBayesian}
\title{Bayesian Contingency Tables}
\usage{
ContingencyTablesBayesian(
  data = NULL,
  version = "1",
  formula = NULL,
  alternative = "twoSided",
  bayesFactorType = "BF10",
  columnOrder = "ascending",
  columns = list(types = list(), value = list()),
  counts = list(types = list(), value = ""),
  countsExpected = FALSE,
  countsObserved = TRUE,
  cramersV = FALSE,
  cramersVCiLevel = 0.95,
  cramersVPlot = FALSE,
  layers = list(),
  marginShowTotals = TRUE,
  oddsRatio = FALSE,
  oddsRatioCiLevel = 0.95,
  percentagesColumn = FALSE,
  percentagesRow = FALSE,
  percentagesTotal = FALSE,
  plotHeight = 240,
  plotWidth = 320,
  posteriorOddsRatioPlot = FALSE,
  posteriorOddsRatioPlotAdditionalInfo = TRUE,
  priorConcentration = 1,
  residualsPearson = FALSE,
  residualsStandardized = FALSE,
  residualsUnstandardized = FALSE,
  rowOrder = "ascending",
  rows = list(types = list(), value = list()),
  samplingModel = "independentMultinomialRowsFixed",
  seed = 1,
  setSeed = FALSE
)
}
\arguments{
\item{countsExpected, }{Show counts expected under the null hypothesis.
Defaults to \code{FALSE}.}

\item{countsObserved, }{Show actual counts from the data.
Defaults to \code{TRUE}.}

\item{marginShowTotals, }{Shows row and column totals.
Defaults to \code{TRUE}.}

\item{oddsRatio, }{Displays the odds ratio for 2×2 tables (ad/bc) with a credible interval.
Defaults to \code{FALSE}.}

\item{oddsRatioCiLevel, }{Range where the true log odds ratio likely falls, given the data.}

\item{percentagesColumn, }{Shows column-wise percentages.
Defaults to \code{FALSE}.}

\item{percentagesRow, }{Shows row-wise percentages.
Defaults to \code{FALSE}.}

\item{percentagesTotal, }{Shows percentages of total.
Defaults to \code{FALSE}.}

\item{posteriorOddsRatioPlot, }{Displays a graphical summary of the log odds ratio.
Defaults to \code{FALSE}.}

\item{priorConcentration, }{Controls how strongly the prior favours equal proportions.}
}
\description{
Bayesian Contingency Tables
}
