% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformedBinomialTestBayesianWrapper.R
\name{InformedBinomialTestBayesian}
\alias{InformedBinomialTestBayesian}
\title{Informed Bayesian Multi-Binomial Test}
\usage{
InformedBinomialTestBayesian(
  data = NULL,
  version = "1",
  bayesFactorType = "BF10",
  bfComparison = "Encompassing",
  bfVsHypothesis = "Model 1",
  bridgeSamples = 1000,
  descriptivesDisplay = "counts",
  descriptivesPlot = FALSE,
  descriptivesTable = FALSE,
  factor = list(types = list(), value = ""),
  includeEncompassingModel = TRUE,
  includeNullModel = TRUE,
  mcmcBurnin = 500,
  mcmcSamples = 5000,
  models = list(list(modelName = "Model 1", syntax = "")),
  plotHeight = 320,
  plotWidth = 480,
  posteriorPlot = FALSE,
  posteriorPlotCiCoverage = 0.95,
  priorCounts = list(list(levels = list(), name = "data 1", values = list()), list(levels
    = list(), name = "data 2", values = list())),
  priorModelProbability = list(list(levels = list(), name = "data 1", values = list())),
  sampleSize = list(types = list(), value = ""),
  seed = 1,
  sequentialAnalysisNumberOfSteps = 10,
  sequentialAnalysisPlot = FALSE,
  sequentialAnalysisPlotType = "bayesFactor",
  setSeed = FALSE,
  successes = list(types = list(), value = "")
)
}
\arguments{
\item{bridgeSamples, }{Max number of samples used to estimate marginal likelihoods for computing Bayes factors.}

\item{descriptivesPlot, }{Shows a plot of observed counts or proportions.
Defaults to \code{FALSE}.}

\item{descriptivesTable, }{Displays a summary table of observed counts or proportions.
Defaults to \code{FALSE}.}

\item{mcmcBurnin, }{Number of initial MCMC samples to discard before analysis begins, allowing the chain to stabilize.}

\item{mcmcSamples, }{Total number of MCMC samples used for estimating the posterior after burn-in.}

\item{posteriorPlot, }{Displays the posterior distribution with credible intervals.
Defaults to \code{FALSE}.}

\item{sequentialAnalysisNumberOfSteps, }{Number of data points at which the Bayes factor is updated during sequential analysis.}

\item{sequentialAnalysisPlot, }{Displays the development of the Bayes factor or posterior probability as the data come in.
Defaults to \code{FALSE}.}
}
\description{
Informed Bayesian Multi-Binomial Test
}
