% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialtestWrapper.R
\name{MultinomialTest}
\alias{MultinomialTest}
\title{Multinomial Test}
\usage{
MultinomialTest(
  data = NULL,
  version = "1",
  count = list(types = list(), value = ""),
  descriptivesPlot = FALSE,
  descriptivesPlotCiLevel = 0.95,
  descriptivesTable = FALSE,
  descriptivesTableCi = FALSE,
  descriptivesTableCiLevel = 0.95,
  descriptivesType = "counts",
  expectedCount = list(types = list(), value = ""),
  factor = list(types = list(), value = ""),
  plotHeight = 320,
  plotWidth = 480,
  testValues = "equal",
  testValuesCustom = list(list(levels = list(), name = "data 1", values = list())),
  vovkSellke = FALSE
)
}
\arguments{
\item{count, }{The variable that contains the count data.}

\item{descriptivesPlot, }{Plots the frequencies and the confidence intervals of the observed counts.
Defaults to \code{FALSE}.}

\item{descriptivesPlotCiLevel, }{Coverage of the confidence intervals in percentages. The default value is 95.}

\item{descriptivesTable, }{Displays the descriptives of the observed and expected counts as well as the confidence intervals of the observed values.
Defaults to \code{FALSE}.}

\item{descriptivesTableCi, }{Coverage of the confidence intervals in percentages. The default value is 95.
Defaults to \code{FALSE}.}

\item{expectedCount, }{If the data includes a variable representing expected cell counts, enter it here; its values define the null hypothesis.}

\item{factor, }{The categorical variable we are interested in.}

\item{vovkSellke, }{An upper bound on how much more likely a p-value is under the alternative hypothesis than under the null.
Defaults to \code{FALSE}.}
}
\description{
The multinomial test allows the user to test whether an observed distribution of cell counts corresponds to an expected distribution.
## Assumptions
- The variable of interest should be categorical.
}
