% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialtestbayesianWrapper.R
\name{MultinomialTestBayesian}
\alias{MultinomialTestBayesian}
\title{Bayesian Multinomial Test}
\usage{
MultinomialTestBayesian(
  data = NULL,
  version = "1",
  bayesFactorType = "BF10",
  count = list(types = list(), value = ""),
  descriptivesPlot = FALSE,
  descriptivesPlotCiLevel = 0.95,
  descriptivesTable = FALSE,
  descriptivesTableCi = FALSE,
  descriptivesTableCiLevel = 0.95,
  descriptivesType = "counts",
  expectedCount = list(types = list(), value = ""),
  factor = list(types = list(), value = ""),
  plotHeight = 320,
  plotWidth = 480,
  priorCounts = list(list(levels = list(), name = "data 1", values = list())),
  testValues = "equal",
  testValuesCustom = list(list(levels = list(), name = "data 1", values = list()))
)
}
\arguments{
\item{count, }{The variable that contains the count data.}

\item{descriptivesPlot, }{Displays the frequency of the reported counts and the corresponding credible intervals for every level of the variable of interest.
Defaults to \code{FALSE}.}

\item{descriptivesPlotCiLevel, }{A credible interval shows the probability that the true effect size lies within certain values. The default credible interval is set at 95%.}

\item{descriptivesTable, }{Displays a table containing the categories of interest, the observed values and the expected values under the specified hypotheses.
Defaults to \code{FALSE}.}

\item{descriptivesTableCi, }{Display central credible intervals. A credible interval shows the probability that the true effect size lies within certain values. The default credible interval is set at 95%.
Defaults to \code{FALSE}.}

\item{expectedCount, }{If the data includes a variable representing expected cell counts, enter it here; its values define the null hypothesis.}

\item{factor, }{The categorical variable we are interested in.}
}
\description{
The Bayesian multinomial test allows the user to test whether an observed distribution of cell counts corresponds to an expected distribution.
## Assumptions
- The variable of interest should be categorical.
}
