% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionloglinearWrapper.R
\name{RegressionLogLinear}
\alias{RegressionLogLinear}
\title{Log-Linear Regression}
\usage{
RegressionLogLinear(
  data = NULL,
  version = "1",
  formula = NULL,
  count = list(types = list(), value = ""),
  factors = list(types = list(), value = list()),
  modelTerms = list(optionKey = "components", types = list(), value = list()),
  plotHeight = 320,
  plotWidth = 480,
  regressionCoefficientsCi = FALSE,
  regressionCoefficientsCiLevel = 0.95,
  regressionCoefficientsEstimates = FALSE,
  vovkSellke = FALSE
)
}
\arguments{
\item{regressionCoefficientsCi, }{Coverage of the confidence intervals in percentages. The default value is 95.
Defaults to \code{FALSE}.}

\item{regressionCoefficientsEstimates, }{Displays estimates of the regression coefficients along with their standard errors, Z-values, and associated p-values.
Defaults to \code{FALSE}.}

\item{vovkSellke, }{An upper bound on how much more likely a p-value is under the alternative hypothesis than under the null.
Defaults to \code{FALSE}.}
}
\description{
Log-linear regression models the logarithm of the dependent variable as a linear combination of independent variables. It is useful for handling skewed data and multiplicative relationships.
}
